Move entities to own folder
