+++ review/internal/api/coupon.go
@@ -1,14 +1,16 @@
 package api
 
 import (
-	. "coupon_service/internal/api/entity"
+	apiEntity "coupon_service/internal/api/entity"
+	"coupon_service/internal/entity"
+
 	"net/http"
 
 	"github.com/gin-gonic/gin"
 )
 
 func (a *API) Apply(c *gin.Context) {
-	apiReq := ApplicationRequest{}
+	apiReq := apiEntity.ApplicationRequest{}
 	if err := c.ShouldBindJSON(&apiReq); err != nil {
 		return
 	}
@@ -21,7 +23,7 @@ func (a *API) Apply(c *gin.Context) {
 }
 
 func (a *API) Create(c *gin.Context) {
-	apiReq := Coupon{}
+	apiReq := entity.Coupon{}
 	if err := c.ShouldBindJSON(&apiReq); err != nil {
 		return
 	}
@@ -33,7 +35,7 @@ func (a *API) Create(c *gin.Context) {
 }
 
 func (a *API) Get(c *gin.Context) {
-	apiReq := CouponRequest{}
+	apiReq := apiEntity.CouponRequest{}
 	if err := c.ShouldBindJSON(&apiReq); err != nil {
 		return
 	}
