+++ review/internal/entity/basket.go
@@ -1,9 +1,5 @@
 package entity
 
-import (
-	_ "github.com/gin-gonic/gin"
-)
-
 type Basket struct {
 	Value                 int
 	AppliedDiscount       int
