+++ review/internal/service/service_test.go
@@ -1,8 +1,8 @@
 package service
 
 import (
+	"coupon_service/internal/entity"
 	"coupon_service/internal/repository/memdb"
-	"coupon_service/internal/service/entity"
 	"reflect"
 	"testing"
 )
